Looping Videos and Remembering Location
After some stress testing, we found that it was good to be able to loop
through videos but remember where a particular video group was up to in
the index.  Otherwise it always started from scratch.  We also found
that the play button was unnecessary and videos are incremented by
pressing the relevant button.